spec/tests/requirements.txt
jsonschema==4.23.0
pyyaml==6.0.2
pytest==8.3.3
rfc3987==1.3.8
